# Use sorted slices to represent HDR histogram