--- a/aggregators/internal/hdrhistogram/hdrhistogram.go
+++ b/aggregators/internal/hdrhistogram/hdrhistogram.go
@@ -124,8 +124,7 @@ func (h *HistogramRepresentation) Buckets() (int64, []int64, []float64) {
 	var totalCount int64
 	var bucketsSeen int
 	iter := h.iterator()
-	// TODO @lahsivjar: Assuming sparse representation, we might be better off
-	// by sorting CountsRep rather than what we do now to avoid sorting.
+	// TODO @lahsivjar: Optimize to use sorted slice representation.
 	for idx := 0; iter.next(); idx++ {
 		if bucketsSeen == h.CountsRep.Len() {
 			break