… aggregators/aggregator_test.go
… aggregators/merger.go
… aggregators/merger_test.go
… go.mod
… proto/aggregation.proto