An alternative to #49. Uses sorted slices to store HDRHistograms. The sorted nature is used to efficiently perform merges.

Benchmark (thanks to #50):

```
name                           old time/op    new time/op    delta
NDJSON/AgentNodeJS-10            33.9ms ± 1%    15.3ms ± 3%  -54.73%  (p=0.000 n=10+10)
NDJSON/AgentPython-10            58.7ms ± 1%    44.1ms ± 2%  -24.76%  (p=0.000 n=9+10)
NDJSON/AgentRuby-10              49.7ms ± 1%    23.7ms ± 2%  -52.36%  (p=0.000 n=9+10)
NDJSON/AgentGo-10                76.2ms ± 1%    26.6ms ± 1%  -65.13%  (p=0.000 n=10+10)
NDJSONParallel/AgentNodeJS-10    33.9ms ± 1%    15.5ms ± 2%  -54.33%  (p=0.000 n=10+10)
NDJSONParallel/AgentPython-10    59.5ms ± 3%    44.1ms ± 2%  -25.82%  (p=0.000 n=10+10)
NDJSONParallel/AgentRuby-10      50.2ms ± 2%    23.4ms ± 3%  -53.42%  (p=0.000 n=10+10)
NDJSONParallel/AgentGo-10        77.0ms ± 2%    26.8ms ± 1%  -65.25%  (p=0.000 n=10+10)

name                           old alloc/op   new alloc/op   delta
NDJSON/AgentNodeJS-10            7.74MB ± 0%    7.58MB ± 0%   -2.04%  (p=0.000 n=10+9)
NDJSON/AgentPython-10            21.4MB ± 1%    20.9MB ± 1%   -2.41%  (p=0.000 n=10+10)
NDJSON/AgentRuby-10              12.0MB ± 0%    11.7MB ± 0%   -1.95%  (p=0.000 n=10+10)
NDJSON/AgentGo-10                15.2MB ± 1%    14.0MB ± 0%   -8.18%  (p=0.000 n=10+10)
NDJSONParallel/AgentNodeJS-10    7.73MB ± 0%    7.59MB ± 0%   -1.76%  (p=0.000 n=10+10)
NDJSONParallel/AgentPython-10    21.4MB ± 1%    20.9MB ± 0%   -2.54%  (p=0.000 n=10+10)
NDJSONParallel/AgentRuby-10      12.0MB ± 0%    11.7MB ± 0%   -1.96%  (p=0.000 n=9+10)
NDJSONParallel/AgentGo-10        15.2MB ± 1%    14.0MB ± 0%   -8.21%  (p=0.000 n=9+9)

name                           old allocs/op  new allocs/op  delta
NDJSON/AgentNodeJS-10             83.4k ± 0%     83.1k ± 0%   -0.36%  (p=0.000 n=10+10)
NDJSON/AgentPython-10              221k ± 1%      223k ± 1%   +0.87%  (p=0.009 n=10+10)
NDJSON/AgentRuby-10                134k ± 0%      134k ± 0%   +0.30%  (p=0.000 n=10+9)
NDJSON/AgentGo-10                  165k ± 0%      162k ± 0%   -1.86%  (p=0.000 n=10+8)
NDJSONParallel/AgentNodeJS-10     83.4k ± 0%     83.1k ± 0%   -0.36%  (p=0.000 n=10+10)
NDJSONParallel/AgentPython-10      222k ± 1%      222k ± 1%     ~     (p=0.165 n=10+10)
NDJSONParallel/AgentRuby-10        134k ± 0%      134k ± 0%   +0.29%  (p=0.000 n=10+7)
```

More benchmarks:

```
name                         old time/op    new time/op    delta
AggregateCombinedMetrics-10    1.64µs ± 2%    1.59µs ± 3%   -2.98%  (p=0.000 n=10+9)
AggregateBatchSerial-10        5.44µs ± 5%    4.76µs ± 2%  -12.48%  (p=0.000 n=10+10)
AggregateBatchParallel-10      5.49µs ± 2%    4.83µs ± 3%  -12.04%  (p=0.000 n=8+10)
CombinedMetricsEncoding-10      982ns ± 2%     960ns ± 0%   -2.27%  (p=0.000 n=9+7)
CombinedMetricsToBatch-10       201µs ± 1%      89µs ± 0%  -55.69%  (p=0.000 n=8+10)
EventToCombinedMetrics-10       312ns ± 1%     310ns ± 0%   -0.73%  (p=0.000 n=10+10)

name                         old alloc/op   new alloc/op   delta
AggregateCombinedMetrics-10    1.49kB ± 0%    1.45kB ± 1%   -2.93%  (p=0.000 n=9+10)
AggregateBatchSerial-10        2.66kB ± 4%    2.54kB ± 4%   -4.56%  (p=0.001 n=10+10)
AggregateBatchParallel-10      2.63kB ± 2%    2.55kB ± 4%   -2.89%  (p=0.003 n=9+10)
CombinedMetricsEncoding-10       654B ± 1%      640B ± 1%   -2.09%  (p=0.000 n=9+9)
CombinedMetricsToBatch-10      4.09kB ± 0%    4.08kB ± 0%   -0.27%  (p=0.000 n=9+10)
EventToCombinedMetrics-10       0.00B          0.00B          ~     (all equal)

name                         old allocs/op  new allocs/op  delta
AggregateCombinedMetrics-10      23.0 ± 0%      21.0 ± 0%   -8.70%  (p=0.000 n=10+10)
AggregateBatchSerial-10          31.6 ± 5%      26.1 ± 4%  -17.29%  (p=0.000 n=9+10)
AggregateBatchParallel-10        31.4 ± 2%      26.4 ± 5%  -15.92%  (p=0.000 n=10+10)
CombinedMetricsEncoding-10       0.00           0.00          ~     (all equal)
CombinedMetricsToBatch-10        95.0 ± 0%      95.0 ± 0%     ~     (all equal)
EventToCombinedMetrics-10        0.00           0.00          ~     (all equal)
```